+++ CHANGELOG.md
@@ -6,4 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add debug module `riscv-dbg` v0.3
+- Add `axi_mem_if` fix-name-conflict
+- Add `debug_subsystem` for RISC-V compliant external debug support (spec v0.13)
+- Add `dm_test`, a debug module testbench
 - Initial development